+++ include/rocksdb/write_buffer_manager.h
@@ -16,17 +16,19 @@
 
 #include <atomic>
 #include <cstddef>
+#include "rocksdb/cache.h"
 
 namespace rocksdb {
 
 class WriteBufferManager {
  public:
-  // _buffer_size = 0 indicates no limit. Memory won't be tracked,
+  // _buffer_size = 0 indicates no limit. Memory won't be capped.
   // memory_usage() won't be valid and ShouldFlush() will always return true.
-  explicit WriteBufferManager(size_t _buffer_size)
-      : buffer_size_(_buffer_size), memory_used_(0) {}
-
-  ~WriteBufferManager() {}
+  // if `cache` is provided, we'll put dummy entries in the cache and cost
+  // the memory allocated to the cache. It can be used even if _buffer_size = 0.
+  explicit WriteBufferManager(size_t _buffer_size,
+                              std::shared_ptr<Cache> cache = {});
+  ~WriteBufferManager();
 
   bool enabled() const { return buffer_size_ != 0; }
 
@@ -34,21 +36,41 @@ class WriteBufferManager {
   size_t memory_usage() const {
     return memory_used_.load(std::memory_order_relaxed);
   }
+  size_t mutable_memtable_memory_usage() const {
+    return memory_active_.load(std::memory_order_relaxed);
+  }
   size_t buffer_size() const { return buffer_size_; }
 
   // Should only be called from write thread
   bool ShouldFlush() const {
-    return enabled() && memory_usage() >= buffer_size();
+    // Flush if memory usage hits a hard limit, or total size that hasn't been
+    // scheduled to free hits a soft limit, which is 7/8 of the hard limit.
+    return enabled() &&
+           (memory_usage() >= buffer_size() ||
+            mutable_memtable_memory_usage() >= buffer_size() / 8 * 7);
   }
 
-  // Should only be called from write thread
   void ReserveMem(size_t mem) {
-    if (enabled()) {
+    if (cache_rep_ != nullptr) {
+      ReserveMemWithCache(mem);
+    } else if (enabled()) {
       memory_used_.fetch_add(mem, std::memory_order_relaxed);
+    }
+    if (enabled()) {
+      memory_active_.fetch_add(mem, std::memory_order_relaxed);
+    }
+  }
+  // We are in the process of freeing `mem` bytes, so it is not considered
+  // when checking the soft limit.
+  void ScheduleFreeMem(size_t mem) {
+    if (enabled()) {
+      memory_active_.fetch_sub(mem, std::memory_order_relaxed);
     }
   }
   void FreeMem(size_t mem) {
-    if (enabled()) {
+    if (cache_rep_ != nullptr) {
+      FreeMemWithCache(mem);
+    } else if (enabled()) {
       memory_used_.fetch_sub(mem, std::memory_order_relaxed);
     }
   }
@@ -56,6 +78,13 @@ class WriteBufferManager {
  private:
   const size_t buffer_size_;
   std::atomic<size_t> memory_used_;
+  // Memory that hasn't been scheduled to free.
+  std::atomic<size_t> memory_active_;
+  struct CacheRep;
+  std::unique_ptr<CacheRep> cache_rep_;
+
+  void ReserveMemWithCache(size_t mem);
+  void FreeMemWithCache(size_t mem);
 
   // No copying allowed
   WriteBufferManager(const WriteBufferManager&) = delete;