+++ util/arena.h
@@ -40,7 +40,8 @@ class Arena : public Allocator {
   // huge_page_size: if 0, don't use huge page TLB. If > 0 (should set to the
   // supported hugepage size of the system), block allocation will try huge
   // page TLB first. If allocation fails, will fall back to normal case.
-  explicit Arena(size_t block_size = kMinBlockSize, size_t huge_page_size = 0);
+  explicit Arena(size_t block_size = kMinBlockSize,
+                 AllocTracker* tracker = nullptr, size_t huge_page_size = 0);
   ~Arena();
 
   char* Allocate(size_t bytes) override;
@@ -114,6 +115,7 @@ class Arena : public Allocator {
 
   // Bytes of memory in blocks allocated so far
   size_t blocks_memory_ = 0;
+  AllocTracker* tracker_;
 };
 
 inline char* Arena::Allocate(size_t bytes) {